revert a wrong warning fix from rev 5257
..otherwise for example volume control is broken

git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5265 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/win32com/VPinMAMEConfig.cpp
@@ -473,7 +473,7 @@ bool RegLoadOpts(HKEY hKey, rc_option *pOpt, char* pszDefault, char* pszValue)
 			fNew = true;
 		}
 		else
-			sprintf(szValue, "%u", dwValue);
+			sprintf(szValue, "%i", (int)dwValue); // needs to always be interpreted as int!
 		break;
 
 	case rc_bool:
